Make it an error to def() an operator multiple times
Summary: Previously, if an operator is defined multiple times, it will be merged anyways. This code changes the behavior to spitting error if an operator is defined more than once.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Make it an error to def() an operator multiple times

Summary: Previously, if an operator is defined multiple times, it will be merged anyways. This code changes the behavior to spitting error if an operator is defined more than once.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9e9d686bf05631e5786177601583d92b8f360718
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39459